Macur/requirements.txt
torch
transformers==4.42.4
trl==0.8.6
rdkit-pypi
tqdm
pandas
numpy
selfies
pyyaml


